+++ naas/api.py
@@ -13,11 +13,10 @@ class Api:
     def __init__(self):
         self.manager = Manager()
         self.get = self.manager.get_prod
+        self.list = self.manager.list_prod
+        self.clear = self.manager.clear_prod
         self.get_output = self.manager.get_output
         self.clear_output = self.manager.clear_output
-        self.get_history = self.manager.get_history
-        self.list_history = self.manager.list_history
-        self.clear_history = self.manager.clear_history
 
     def currents(self, raw=False):
         json_data = self.manager.get_naas()
@@ -42,15 +41,14 @@ class Api:
         if token is None:
             token = os.urandom(30).hex()
         url = self.manager.proxy_url(self.role, token)
-        if not self.manager.notebook_path():
-            print("No add done you are in already in naas folder\n")
+        if self.manager.is_production():
+            print("No add done you are in production\n")
             return url
         self.manager.add_prod(
             {"type": self.role, "path": current_file, "params": params, "value": token},
             debug,
         )
-        print("👌 Well done! Your Notebook has been sent to production folder.\n")
-        print(f"🔑 You can run this notebook remotely with: {url} \n")
+        print("👌 Well done! Your Notebook has been sent to production.\n")
         self.manager.copy_url(url)
         print(
             'PS: to remove the "Notebook as API" feature, just replace .add by .delete'
@@ -66,10 +64,14 @@ class Api:
         )
 
     def respond_file(self, path):
-        naas_type = mimetypes.guess_type(os.path.abspath(path))[0]
-        data = {"path": os.path.abspath(path)}
+        abs_path = os.path.abspath(path)
+        naas_type = mimetypes.guess_type(abs_path)[0]
         display(Markdown("Response Set as File, preview below: "))
-        display(JSON(data, metadata={"naas_api": True, "naas_type": naas_type}))
+        display(
+            JSON(
+                {"path": abs_path}, metadata={"naas_api": True, "naas_type": naas_type}
+            )
+        )
 
     def respond_html(self, data):
         display(Markdown("Response Set as HTML, preview below: "))
@@ -79,11 +81,9 @@ class Api:
         display(Markdown("Response Set as JSON, preview below: "))
         display(JSON(data, metadata={"naas_api": True}))
 
-    def respond_image(self, data, filename, mimetype):
+    def respond_image(self, data, filename):
         display(Markdown("Response Set as IMAGE, preview below: "))
-        display(
-            Image(data, filename=filename, format=mimetype, metadata={"naas_api": True})
-        )
+        display(Image(data, filename=filename, metadata={"naas_api": True}))
 
     def respond_svg(self, data):
         display(Markdown("Response Set as SVG, preview below: "))
@@ -99,38 +99,12 @@ class Api:
         display(HTML(data.to_html(), metadata={"naas_api": True, "naas_type": "csv"}))
 
     def delete(self, path=None, all=False, debug=False):
-        if not self.manager.notebook_path():
-            print("No delete done you are in already in naas folder\n")
+        if self.manager.is_production():
+            print("No delete done you are in production\n")
             return
         current_file = self.manager.get_path(path)
         self.manager.del_prod({"type": self.role, "path": current_file}, debug)
-        print("🗑 Done! Your Notebook has been remove from production folder.\n")
+        print("🗑 Done! Your Notebook has been remove from production.\n")
         if all is True:
-            self.clear_history(path)
+            self.clear(path)
             self.clear_output(path)
-
-    def help(self):
-        print(f"=== {type(self).__name__} === \n")
-        print(
-            f".add(path, params) => add path to the prod {type(self).__name__} server\n"
-        )
-        print(
-            f".delete(path) => delete path to the prod {type(self).__name__} server\n"
-        )
-        print(
-            ".clear_history(path, histonumber) => clear history, history number and path are optionel, \
-                if you don't provide them it will erase full history of current file \n"
-        )
-        print(
-            ".list_history(path) => list history, of a path or if not provided the current file \n"
-        )
-        print(
-            ".get_history(histonumber, path) => get history file, of a path or if not provided the current file \n"
-        )
-        print(
-            ".get(path) => get current prod file of a path, or if not provided the current file \n"
-        )
-        print(f".currents() => get current list of {type(self).__name__} prod file\n")
-        print(
-            f".current(raw=True) => get json current list of {type(self).__name__} prod file\n"
-        )