+++ naas/dependency.py
@@ -9,9 +9,8 @@ class Dependency:
     def __init__(self):
         self.manager = Manager()
         self.get = self.manager.get_prod
-        self.get_history = self.manager.get_history
-        self.list_history = self.manager.list_history
-        self.clear_history = self.manager.clear_history
+        self.list = self.manager.list_prod
+        self.clear = self.manager.clear_prod
 
     def currents(self, raw=False):
         json_data = self.manager.get_naas()
@@ -25,8 +24,8 @@ class Dependency:
                     print(f"File ==> {item['path']}")
 
     def add(self, path=None, debug=False):
-        if not self.manager.notebook_path():
-            print("No add done you are in already in naas folder\n")
+        if self.manager.is_production():
+            print("No add done you are in production\n")
             return
         current_file = self.manager.get_path(path)
         self.manager.add_prod(
@@ -38,39 +37,15 @@ class Dependency:
             },
             debug,
         )
-        print("👌 Well done! Your Dependency has been sent to production folder. \n")
+        print("👌 Well done! Your Dependency has been sent to production. \n")
         print('PS: to remove the "Dependency" feature, just replace .add by .delete')
 
     def delete(self, path=None, all=False, debug=False):
-        if not self.manager.notebook_path():
-            print("No delete done you are in already in naas folder\n")
+        if self.manager.is_production():
+            print("No delete done you are in production\n")
             return
         current_file = self.manager.get_path(path)
         self.manager.del_prod({"type": self.role, "path": current_file}, debug)
-        print("🗑 Done! Your Dependency has been remove from production folder.\n")
+        print("🗑 Done! Your Dependency has been remove from production.\n")
         if all is True:
-            self.clear_history(path)
-
-    def help(self):
-        print(f"=== {type(self).__name__} === \n")
-        print(".add(path) => add path to the prod server\n")
-        print(
-            f".delete(path) => delete path to the prod {type(self).__name__} server\n"
-        )
-        print(
-            ".clear_history(path, histonumber) => clear history, history number and path are optionel, \
-                if you don't provide them it will erase full history of current file \n"
-        )
-        print(
-            ".list_history(path) => list history, of a path or if not provided the current file \n"
-        )
-        print(
-            ".get_history(path, histonumber) => get history file, of a path or if not provided the current file \n"
-        )
-        print(
-            ".get(path) => get current prod file of a path, or if not provided the current file \n"
-        )
-        print(f".currents() => get current list of {type(self).__name__} prod file\n")
-        print(
-            f".current(raw=True) => get json current list of {type(self).__name__} prod file\n"
-        )
+            self.clear(path)