+++ naas/runner/controllers/logs.py
@@ -1,7 +1,8 @@
 from sanic.views import HTTPMethodView
-from sanic.response import json, file
+from sanic import response
 from naas.types import t_asset
 import uuid
+import json
 
 
 class LogsController(HTTPMethodView):
@@ -14,14 +15,17 @@ class LogsController(HTTPMethodView):
     async def get(self, request):
         as_file = request.args.get("as_file", False)
         if as_file:
-            return await file(self.__logger.get_file_path(), filename="logs.csv")
+            return await response.file(
+                self.__logger.get_file_path(), filename="logs.csv"
+            )
         else:
             uid = str(uuid.uuid4())
             limit = int(request.args.get("limit", 0))
             skip = int(request.args.get("skip", 0))
             search = str(request.args.get("search", ""))
-            filters = list(request.args.get("filters", []))
-            logs = self.__logger.list(uid, skip, limit, search, filters)
+            sort = list(json.loads(request.args.get("sort", "[]")))
+            filters = list(json.loads(request.args.get("filters", "[]")))
+            logs = self.__logger.list(uid, skip, limit, search, filters, sort)
             self.__logger.info(
                 {
                     "id": uid,
@@ -32,6 +36,7 @@ class LogsController(HTTPMethodView):
                     "limit": limit,
                     "search": search,
                     "filters": filters,
+                    "sort": sort,
                 }
             )
-            return json(logs)
+            return response.json(logs)