+++ naas/runner/logger.py
@@ -97,6 +97,7 @@ class Logger:
         limit: int = 0,
         search: str = "",
         filters: list = [],
+        sort: list = [],
     ):
         df = None
         try:
@@ -110,6 +111,11 @@ class Logger:
             df = pd.concat([df1, df], axis=1, sort=False)
             if len(filters) > 0:
                 df = df[df.type.isin(filters)]
+            if len(sort) > 0:
+                for query in sort:
+                    field = [query["field"]]
+                    ascending = False if query["type"] == "desc" else True
+                    df = df.sort_values(by=field, ascending=ascending)
             total_records = len(df.index)
             if search and search != "":
                 idx = df.apply(
@@ -123,10 +129,11 @@ class Logger:
                 df = df[:limit]
             df = df.reset_index()
             return {
+                "uid": uid,
                 "data": json.loads(df.to_json(orient="records")),
                 "totalRecords": total_records,
             }
         except Exception as e:
             tb = traceback.format_exc()
             print("list logs", e, tb)
-            return {"data": [], "totalRecords": 0}
+            return {"uid": uid, "data": [], "totalRecords": 0}