+++ naas/runner/proxy.py
@@ -27,13 +27,6 @@ def escape_docker(s):
     )
 
 
-def get_status_server():
-    req = requests.get(url=os.environ.get("PUBLIC_PROXY_API", ""))
-    req.raise_for_status()
-    jsn = req.json()
-    return jsn
-
-
 def encode_proxy_url(token=""):
     client = os.environ.get("JUPYTERHUB_USER", "")
     base_public_url = os.environ.get("PUBLIC_PROXY_API", "")
@@ -42,3 +35,61 @@ def encode_proxy_url(token=""):
     base64_bytes = b64encode(message_bytes)
     username_base64 = base64_bytes.decode("ascii")
     return f"{base_public_url}/{username_base64}/{token}"
+
+
+class Domain:
+
+    base_notif_url = os.environ.get("PUBLIC_PROXY_API", None)
+    headers = None
+
+    def __init__(self):
+        self.headers = {
+            "Authorization": f'token {os.environ.get("JUPYTERHUB_API_TOKEN", None)}'
+        }
+
+    def status(self):
+        req = requests.get(url=f"{self.base_notif_url}/status")
+        req.raise_for_status()
+        jsn = req.json()
+        return jsn
+
+    def add(self, domain, url=None):
+        token = None
+        endpoint = None
+        if url:
+            list_url = url.split("/")
+            token = list_url.pop()
+            endpoint = list_url.pop()
+        if "://" in domain:
+            clean_domain = domain.split("://")[1]
+        else:
+            clean_domain = domain
+        data = {"domain": clean_domain, "endpoint": endpoint, "token": token}
+        req = requests.post(
+            url=f"{self.base_notif_url}/proxy", headers=self.headers, json=data
+        )
+        req.raise_for_status()
+        new_url = f"https://{clean_domain}"
+        if token:
+            new_url = f"{new_url}/{endpoint}/{token}"
+        return new_url
+
+    def get(self, domain):
+        req = requests.get(
+            url=f"{self.base_notif_url}/proxy",
+            headers=self.headers,
+            json={"domain": domain},
+        )
+        req.raise_for_status()
+        jsn = req.json()
+        return jsn
+
+    def delete(self, domain):
+        req = requests.delete(
+            url=f"{self.base_notif_url}/proxy",
+            headers=self.headers,
+            json={"domain": domain},
+        )
+        req.raise_for_status()
+        jsn = req.json()
+        return jsn