+++ naas/runner/scheduler.py
@@ -2,6 +2,7 @@ from naas.types import t_scheduler, t_start, t_main, t_health, t_error
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 import apscheduler.schedulers.base
 import traceback
+import requests
 import datetime
 import asyncio
 import pycron
@@ -47,16 +48,21 @@ class Scheduler:
                 self.__scheduler.start()
                 uid = str(uuid.uuid4())
                 self.__logger.info(
-                    {"id": uid, "type": t_main, "status": "start SCHEDULER"}
+                    {
+                        "id": uid,
+                        "type": t_main,
+                        "filepath": "sheduler",
+                        "status": "start SCHEDULER",
+                    }
                 )
 
     async def __scheduler_greenlet(self, main_uid, current_time, task):
+        value = task.get("value", None)
+        current_type = task.get("type", None)
+        file_filepath = task.get("path", None)
+        params = task.get("params", dict())
+        uid = str(uuid.uuid4())
         try:
-            value = task.get("value", None)
-            current_type = task.get("type", None)
-            file_filepath = task.get("path")
-            params = task.get("params", dict())
-            uid = str(uuid.uuid4())
             running = await self.__jobs.is_running(uid, file_filepath, current_type)
             if (
                 current_type == t_scheduler
@@ -109,6 +115,33 @@ class Scheduler:
                         "duration": res.get("duration"),
                     }
                 )
+                next_url = params.get("next_url", None)
+                if next_url is not None:
+                    if "http" not in next_url:
+                        next_url = f"{self.__api_internal}{next_url}"
+                    self.__logger.info(
+                        {
+                            "id": uid,
+                            "type": t_scheduler,
+                            "filepath": file_filepath,
+                            "status": "next_url",
+                            "url": next_url,
+                        }
+                    )
+                    try:
+                        req = requests.get(url=next_url)
+                        req.raise_for_status()
+                    except requests.exceptions.RequestException as e:
+                        self.__logger.error(
+                            {
+                                "id": main_uid,
+                                "type": t_scheduler,
+                                "status": t_error,
+                                "filepath": file_filepath,
+                                "error": "Error in next_url",
+                                "trace": str(e),
+                            }
+                        )
                 await self.__jobs.update(
                     uid,
                     file_filepath,
@@ -125,6 +158,7 @@ class Scheduler:
                     "id": main_uid,
                     "type": t_scheduler,
                     "status": t_error,
+                    "filepath": file_filepath,
                     "error": "Unknow error",
                     "trace": str(tb),
                 }
@@ -133,11 +167,18 @@ class Scheduler:
     async def __scheduler_function(self):
         # Create unique for scheduling step (one step every minute)
         main_uid = str(uuid.uuid4())
+        all_start_time = time.time()
         try:
             current_time = datetime.datetime.now()
-            all_start_time = time.time()
             # Write self.__scheduler init info in self.__logger.write
-            self.__logger.info({"id": main_uid, "type": t_scheduler, "status": t_start})
+            self.__logger.info(
+                {
+                    "id": main_uid,
+                    "type": t_scheduler,
+                    "filepath": "scheduler",
+                    "status": t_start,
+                }
+            )
             jobs = await self.__jobs.list(main_uid)
             await asyncio.gather(
                 *[
@@ -150,6 +191,7 @@ class Scheduler:
                 {
                     "id": main_uid,
                     "type": t_scheduler,
+                    "filepath": "scheduler",
                     "status": t_health,
                     "duration": duration_total,
                 }
@@ -162,6 +204,7 @@ class Scheduler:
                     "id": main_uid,
                     "type": t_scheduler,
                     "status": t_error,
+                    "filepath": "scheduler",
                     "duration": duration_total,
                     "error": str(e),
                     "trace": tb,
@@ -175,6 +218,7 @@ class Scheduler:
                     "id": main_uid,
                     "type": t_scheduler,
                     "status": t_error,
+                    "filepath": "scheduler",
                     "duration": duration_total,
                     "error": "Unknow error",
                     "trace": tb,