+++ naas/scheduler.py
@@ -1,6 +1,7 @@
 from .types import t_scheduler
 from .manager import Manager
 import pretty_cron
+import requests
 import pycron
 
 
@@ -10,12 +11,32 @@ class Scheduler:
 
     def __init__(self):
         self.manager = Manager()
-        self.get = self.manager.get_prod
         self.get_output = self.manager.get_output
         self.clear_output = self.manager.clear_output
-        self.get_history = self.manager.get_history
-        self.list_history = self.manager.list_history
-        self.clear_history = self.manager.clear_history
+        self.list = self.manager.list_prod
+        self.clear = self.manager.clear_prod
+        self.get = self.manager.get_prod
+
+    def status(self):
+        req = requests.get(url=f"{self.manager.naas_api}/scheduler")
+        req.raise_for_status()
+        jsn = req.json()
+        print(jsn)
+        return jsn
+
+    def pause(self):
+        req = requests.get(url=f"{self.manager.naas_api}/scheduler/pause")
+        req.raise_for_status()
+        jsn = req.json()
+        print(jsn)
+        return jsn
+
+    def resume(self):
+        req = requests.get(url=f"{self.manager.naas_api}/scheduler/resume")
+        req.raise_for_status()
+        jsn = req.json()
+        print(jsn)
+        return jsn
 
     def currents(self, raw=False):
         json_data = self.manager.get_naas()
@@ -41,8 +62,8 @@ class Scheduler:
         return res
 
     def add(self, path=None, recurrence=None, params=None, debug=False):
-        if not self.manager.notebook_path():
-            print("No add done you are in already in naas folder\n")
+        if self.manager.is_production():
+            print("No add done you are in production\n")
             return
         if not recurrence:
             print("No recurrence provided\n")
@@ -61,45 +82,19 @@ class Scheduler:
             debug,
         )
         cron_string = pretty_cron.prettify_cron(recurrence)
-        print("👌 Well done! Your Notebook has been sent to production folder. \n")
+        print("👌 Well done! Your Notebook has been sent to production. \n")
         print(
             f'⏰ It will be scheduled "{cron_string}" (more on the syntax on https://crontab.guru/).\n'
         )
         print('Ps: to remove the "Scheduler", just replace .add by .delete')
 
     def delete(self, path=None, all=False, debug=False):
-        if not self.manager.notebook_path():
-            print("No delete done you are in already in naas folder\n")
+        if self.manager.is_production():
+            print("No delete done you are in production\n")
             return
         current_file = self.manager.get_path(path)
         self.manager.del_prod({"type": self.role, "path": current_file}, debug)
-        print("🗑 Done! Your Scheduler has been remove from production folder.\n")
+        print("🗑 Done! Your Scheduler has been remove from production.\n")
         if all is True:
-            self.clear_history(path)
+            self.clear(path)
             self.clear_output(path)
-
-    def help(self):
-        print(f"=== {type(self).__name__} === \n")
-        print(
-            f".add(path, params) => add path to the prod {type(self).__name__} server\n"
-        )
-        print(
-            f".delete(path) => delete path to the prod {type(self).__name__} server\n"
-        )
-        print(
-            ".clear_history(histonumber, path) => clear history, history number and path are optionel, \
-                if you don't provide them it will erase full history of current file \n"
-        )
-        print(
-            ".list_history(path) => list history, of a path or if not provided the current file \n"
-        )
-        print(
-            ".get_history(histonumber, path) => get history file, of a path or if not provided the current file \n"
-        )
-        print(
-            ".get(path) => get current prod file of a path, or if not provided the current file \n"
-        )
-        print(f".currents() => get current list of {type(self).__name__} prod file\n")
-        print(
-            f".current(raw=True) => get json current list of {type(self).__name__} prod file\n"
-        )